Refactored backend specific nest level configs to respective relevant header files
