+++ core/src/OpenMPTarget/Kokkos_OpenMPTarget_Instance.hpp
@@ -88,6 +88,11 @@ class OpenMPTargetInternal {
   uint32_t m_instance_id = Kokkos::Tools::Experimental::Impl::idForInstance<
       Kokkos::Experimental::OpenMPTarget>(reinterpret_cast<uintptr_t>(this));
 };
+
+template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
+struct ThreadAndVectorNestLevel<Rank, OpenMPTarget, ThreadAndVector>
+    : HostBasedNestLevel<Rank, ThreadAndVector> {};
+
 }  // Namespace Impl
 }  // Namespace Experimental
 }  // Namespace Kokkos
