+++ core/src/impl/Kokkos_TeamMDPolicy.hpp
@@ -139,75 +139,6 @@ struct AcceleratorBasedNestLevel {
   static constexpr int invalid = -2;
 };
 
-// ThreadAndVectorNestLevel determines on which nested level parallelization
-// happens.
-//   - Rank is Kokkos::Rank<TotalNestLevel, Iter>
-//     - TotalNestLevel is the total number of loop nests
-//     - Iter is whether to go forward or backward through ranks (i.e. the
-//       iteration order for MDRangePolicy)
-//   - ThreadAndVector determines whether both vector and thread parallelism is
-//   in use
-template <typename Rank, typename ExecSpace,
-          TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel;
-
-#ifdef KOKKOS_ENABLE_SERIAL
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, Serial, ThreadAndVector>
-    : HostBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-#ifdef KOKKOS_ENABLE_CUDA
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, Cuda, ThreadAndVector>
-    : AcceleratorBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-#ifdef KOKKOS_ENABLE_HIP
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, HIP, ThreadAndVector>
-    : AcceleratorBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-// TODO check par_rt and par_rv validity
-#ifdef KOKKOS_ENABLE_OPENMPTARGET
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, Experimental::OpenMPTarget,
-                                ThreadAndVector>
-    : HostBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-#ifdef KOKKOS_ENABLE_OPENMP
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, OpenMP, ThreadAndVector>
-    : HostBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-#ifdef KOKKOS_ENABLE_THREADS
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, Threads, ThreadAndVector>
-    : HostBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-// TODO check par_rt and par_rv validity
-#ifdef KOKKOS_ENABLE_SYCL
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, Experimental::SYCL, ThreadAndVector>
-    : HostBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-#ifdef KOKKOS_ENABLE_HPX
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, Experimental::HPX, ThreadAndVector>
-    : HostBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
-#ifdef KOKKOS_ENABLE_OPENACC
-template <typename Rank, TeamMDRangeThreadAndVector ThreadAndVector>
-struct ThreadAndVectorNestLevel<Rank, Experimental::OpenACC, ThreadAndVector>
-    : HostBasedNestLevel<Rank, ThreadAndVector> {};
-#endif
-
 template <typename TeamHandle>
 KOKKOS_INLINE_FUNCTION auto nested_policy(
     TeamMDRangeMode<TeamMDRangeLastNestLevel::NotLastNestLevel,
